Some fixes + docstrings
Updated project structure.
- new modules ('eval' for evaluating operator and boundaries, 'losses' for all losses).
Minor improvements due to new improvements.
