+++ examples/example_Burgers_paper.py
@@ -20,7 +20,7 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname( __file__ ), '..'))
 
 from tedeous.input_preprocessing import Equation
 from tedeous.solver import Solver
-from tedeous.metrics import Solution
+from tedeous.solution import Solution
 
 
 def solver_burgers(grid_res, cache, optimizer, iterations):
