+++ examples/example_KdV_matrix.py
@@ -21,7 +21,7 @@ sys.path.append('../')
 
 from tedeous.input_preprocessing import Equation
 from tedeous.solver import Solver, grid_format_prepare, Plots
-from tedeous.metrics import Solution
+from tedeous.solution import Solution
 from tedeous.device import solver_device
