+++ examples/example_ODE_Legendre_autograd.py
@@ -26,7 +26,7 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname( __file__ ), '..'))
 
 from tedeous.input_preprocessing import Equation
 from tedeous.solver import Solver, grid_format_prepare
-from tedeous.metrics import Solution
+from tedeous.solution import Solution
 from tedeous.device import solver_device
