+++ tedeous/eval.py
@@ -1,18 +1,9 @@
-from __future__ import annotations
-
 import torch
-import numpy as np
-from copy import deepcopy
 from typing import Tuple, Union
 
 from tedeous.points_type import Points_type
 from tedeous.derivative import Derivative
 from tedeous.device import device_type, check_device
-import tedeous.input_preprocessing
-from tedeous.utils import *
-
-flatten_list = lambda t: [item for sublist in t for item in sublist]
-
 
 def integration(func: torch.tensor, grid, pow: Union[int, float] = 2) -> Union[Tuple[float, float], Tuple[list, torch.Tensor]]:
     """
@@ -148,6 +139,7 @@ class Operator():
             return self.pde_compute()
         else:
             return self.weak_pde_compute(self.weak_form)
+
 
 class Bounds():
     """
@@ -295,227 +287,3 @@ class Bounds():
         true_bval_dict.update((key, torch.cat(value)) for key, value in true_bval_dict.items())
 
         return bval_dict, true_bval_dict
-
-class Losses():
-    """
-    Class which contains all losses.
-    """
-    def __init__(self, operator, bval, true_bval, lambda_op, lambda_bound, mode, weak_form, n_t):
-        self.operator = operator
-        self.mode = mode
-        self.weak_form = weak_form
-        self.bval = bval
-        self.true_bval = true_bval
-        self.lambda_bound = tedeous.input_preprocessing.lambda_prepare(bval, lambda_bound)
-        self.lambda_op = tedeous.input_preprocessing.op_lambda_prepare(operator, lambda_op)
-        self.n_t = n_t
-    def loss_op(self):
-        loss_operator = 0
-        for eq in self.operator:
-            if self.weak_form != None and self.weak_form != []:
-                loss_operator += self.lambda_op[eq] * torch.sum(self.operator[eq])
-            else:
-                loss_operator += self.lambda_op[eq] * torch.sum(torch.mean((self.operator[eq]) ** 2, 0))
-        return loss_operator
-
-    def loss_bcs(self) -> torch.Tensor:
-        """
-        Computes boundary loss for corresponding type.
-
-        Returns:
-            boundary loss
-        """
-        loss_bnd = 0
-        for bcs_type in self.bval:
-            loss_bnd += self.lambda_bound[bcs_type] * torch.mean((self.bval[bcs_type] - self.true_bval[bcs_type]) ** 2)
-        return loss_bnd
-
-    def default_loss(self) -> torch.Tensor:
-        """
-        Computes l2 loss.
-
-        Args:
-            lambda_bound: an arbitrary chosen constant, influence only convergence speed.
-        Returns:
-            model loss.
-        """
-        if self.bval == None:
-            return torch.sum(torch.mean((self.operator) ** 2, 0))
-
-        if self.mode == 'mat':
-            loss = torch.mean((self.operator) ** 2) + self.loss_bcs()
-        else:
-            loss = self.loss_op() + self.loss_bcs()
-        return loss
-
-    def causal_loss(self, tol: float = 0) -> torch.Tensor:
-        """
-        Computes causal loss, which is caclulated with weights matrix:
-        W = exp(-tol*(Loss_i)) where Loss_i is sum of the L2 loss from 0
-        to t_i moment of time. This loss function should be used when one
-        of the DE independent parameter is time.
-
-        Args:
-            lambda_bound: an arbitrary chosen constant, influence only convergence speed.
-            tol: float constant, influences on error penalty.
-        Returns:
-            model loss.
-        """
-        op = self.operator ** 2
-
-        if self.bval == None:
-            return torch.sum(torch.mean((op) ** 2, 0))
-
-        res = torch.sum(op, dim=1).reshape(self.n_t, -1)
-        res = torch.mean(res, axis=1).reshape(self.n_t, 1)
-        M = torch.triu(torch.ones((self.n_t, self.n_t)), diagonal=1).T
-        with torch.no_grad():
-            W = torch.exp(- tol * (M @ res))
-
-        loss = torch.mean(W * res) + self.loss_bcs()
-
-        return loss
-
-    def weak_loss(self) -> torch.Tensor:
-        """
-        Weak solution of O/PDE problem.
-
-        Args:
-            weak_form: list of basis functions.
-            lambda_bound: const regularization parameter.
-        Returns:
-            model loss.
-        """
-        if self.bval == None:
-            return torch.sum(self.operator)
-
-        # we apply no  boundary conditions operators if they are all None
-
-        loss = self.loss_op() + self.loss_bcs()
-
-        return loss
-
-    def compute(self, tol: float = 0) -> \
-        Union[Losses.default_loss, Losses.weak_loss, Losses.casual_loss]:
-            """
-            Setting the required loss calculation method.
-
-            Args:
-                lambda_bound: an arbitrary chosen constant, influence only convergence speed.
-                weak_form: list of basis functions.
-                tol: float constant, influences on error penalty.
-
-            Returns:
-                A given calculation method.
-            """
-            if self.mode == 'mat' or self.mode == 'autograd':
-                if self.bval == None:
-                    print('No bconds is not possible, returning infinite loss')
-                    return np.inf
-
-            if self.weak_form != None and self.weak_form != []:
-                return self.weak_loss()
-            elif tol != 0:
-                return self.casual_loss(tol=tol)
-            else:
-                return self.default_loss()
-
-class Solution():
-    """
-    class for different loss functions calculation.
-    """
-    def __init__(self, grid: torch.Tensor, equal_cls: Union[tedeous.input_preprocessing.Equation_NN,
-                                                            tedeous.input_preprocessing.Equation_mat,
-                                                            tedeous.input_preprocessing.Equation_autograd],
-                 model: Union[torch.nn.Sequential, torch.Tensor], mode: str, weak_form, lambda_operator, lambda_bound):
-        self.grid = check_device(grid)
-        if mode == 'NN':
-            sorted_grid = Points_type(self.grid).grid_sort()
-            self.n_t = len(sorted_grid['central'][:, 0].unique())
-        elif mode == 'autograd':
-            self.n_t = len(self.grid[:, 0].unique())
-        elif mode == 'mat':
-            self.n_t = grid.shape[1]
-        equal_copy = deepcopy(equal_cls)
-        self.prepared_operator = equal_copy.operator_prepare()
-        self.prepared_bconds = equal_copy.bnd_prepare()
-        self.model = model.to(device_type())
-        self.mode = mode
-        self.weak_form = weak_form
-        self.lambda_operator = lambda_operator
-        self.lambda_bound = lambda_bound
-        self.operator = Operator(self.grid, self.prepared_operator, self.model,
-                                   self.mode, weak_form)
-        self.boundary = Bounds(self.grid, self.prepared_bconds, self.model,
-                                   self.mode, weak_form)
-        self.op_list = []
-        self.bcs_list = []
-        self.loss_list = []
-
-    @staticmethod
-    def samples_count(second_order_interactions, sampling_N, op, bval):
-        num_of_eq = len(op)
-        grid_len = len(list(op.values())[0].cpu())
-        bval_len = sum((map(lambda x: len(x), bval.values())))
-
-        sampling_D = num_of_eq * grid_len + bval_len
-
-        if second_order_interactions:
-            sampling_amount = sampling_N * (2 * sampling_D + 2)
-        else:
-            sampling_amount = sampling_N * (sampling_D + 2)
-        return sampling_amount, sampling_D
-
-    def evaluate(self,
-                 second_order_interactions: bool = True,
-                 sampling_N: int = 1,
-                 lambda_update: bool = False ,
-                 tol: float = 0)-> torch.Tensor:
-        """
-        Computes loss.
-
-        Args:
-            second_order_interactions: optimizer iteration (serves only for computing adaptive lambdas).
-            sampling_N: parameter for accumulation of solutions (op, bcs). The more sampling_N, the more accurate the estimation of the variance.
-            lambda_update: update lambda or not.
-            tol: float constant, influences on error penalty.
-
-        Returns:
-            loss
-        """
-
-        op = self.operator.operator_compute()
-        op = tensor_to_dict(op)
-        bval, true_bval = self.boundary.apply_bcs()
-
-        loss_cls = Losses(operator=op, bval=bval, true_bval=true_bval, lambda_op=self.lambda_operator,
-                          lambda_bound=self.lambda_bound, mode=self.mode,
-                          weak_form=self.weak_form, n_t=self.n_t)
-
-        loss = loss_cls.compute(tol)
-
-        if lambda_update:
-            op, bcs, op_length, bval_length = wrap(op, bval, true_bval)
-            self.op_list.append(list_to_vector(op.values()).cpu().detach().numpy())
-            self.bcs_list.append(list_to_vector(bcs.values()).cpu().detach().numpy())
-            self.loss_list.append(float(loss.item()))
-            sampling_amount, sampling_D = self.samples_count(second_order_interactions = second_order_interactions,
-                                                             sampling_N = sampling_N,
-                                                             op=op, bval = bcs)
-            # print(len(self.op_list))
-            if len(self.op_list) == sampling_amount:
-                self.lambda_operator, self.lambda_bound = Lambda(self.op_list, self.bcs_list,
-                                                                 self.loss_list, second_order_interactions).update(op_length=op_length,
-                                                                                                                           bcs_length=bval_length,
-                                                                                                                           sampling_D=sampling_D)
-                self.op_list.clear()
-                self.bcs_list.clear()
-                self.loss_list.clear()
-                lambda_print(self.lambda_operator)
-                lambda_print(self.lambda_bound)
-                # print(self.lambda_operator, self.lambda_bound)
-
-        return loss
-def lambda_print(dct):
-    for key in dct.keys():
-        print('lambda_{}: {}'.format(key, dct[key]))
